+++ src/app/recipe/page.tsx
@@ -1,0 +1,9 @@
+import Recipe from "@/components/recipe/Recipe";
+
+export default async function RecipePage() {
+  return (
+    <>
+      <Recipe />
+    </>
+  );
+}
